库更新D240807

+++ lib/include/IA32/string.h
@@ -6,8 +6,8 @@
 #ifndef _STRING_H_
 #define _STRING_H_
 
-#include <IA32/stddef.h>
-void memcpy(void*dect,void*src,u32 len);
-void memset(void*dect,u32 value,u32 len);
+#include <ia32/stddef.h>
+void memcpy(void*dect,void*src,size_t len);
+void memset(void*dect,u32 value,size_t len);
 
 #endif
